% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_Correlation.R
\name{Numeric_Correlation}
\alias{Numeric_Correlation}
\title{A NBA Data Function}
\usage{
Numeric_Correlation(x)
}
\description{
This function allows you to creat correlation matrix for all numeric variables in NBA data for a specific year.
}
\examples{
Numeric_Correlation(1991)
}
\keyword{correlation}
\keyword{matrix,}
\keyword{numeric}
\keyword{variables}
